% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_filter.R
\name{add_filter}
\alias{add_filter}
\title{Add a filter to a table of filters}
\usage{
add_filter(filters, colname, data)
}
\arguments{
\item{filters}{A data frame of filters. See \link{filters_init}.}

\item{colname}{A string specifying the column to filter.}

\item{data}{The data that is to be filtered.}
}
\value{
\code{filters} with an added row, provided that \code{colname} is valid.
}
\description{
Create a filter for a data frame, and add it to a table of filters, where
each filter is represented by a row. The filter is only added if the column
you specify to filter is valid.
}
\details{
The function will either create a character filter or a numeric filter,
depending on the type of the column. See \link{filters_init} for how these two
filters work.
}
\examples{
data <- tibble::tibble(
  x = 1:10,
  y = "a"
)

add_filter(filters_init, "x", data)
add_filter(filters_init, "y", data)
add_filter(filters_init, "z", data)

}
\seealso{
\itemize{
\item \link{filters_init} for the filters table.
\item \code{\link[=edit_filter]{edit_filter()}} and \code{\link[=remove_filter]{remove_filter()}} to further edit the filters table.
\item \code{\link[=apply_filters]{apply_filters()}} to apply the filters you have created to your data.
}
}
