% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_filters.R
\name{apply_filters}
\alias{apply_filters}
\title{Apply a table of filters to a data frame.}
\usage{
apply_filters(df, filters)
}
\arguments{
\item{df}{The data frame to filter. See \link{filters_init}.}

\item{filters}{A data frame of filters.}
}
\value{
A filtered version of \code{df} (a subset of the rows), or \code{NULL} if all rows are
removed.
}
\description{
Use a data frame, where each row represents a filter, to filter another data
frame using the value of its columns.
}
\examples{
data <- tibble::tibble(
  x = 1:10,
  y = "a"
)

filters <- add_filter(filters_init, "x", data)
filters <- edit_filter(filters, 1, min = 1, max = 5)
apply_filters(data, filters)

}
\seealso{
\itemize{
\item \link{filters_init} for the filters table.
\item \code{\link[=add_filter]{add_filter()}}, \code{\link[=edit_filter]{edit_filter()}} and \code{\link[=remove_filter]{remove_filter()}} to manipulate the
filters table.
}
}
