% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_scores.R
\name{apply_scores}
\alias{apply_scores}
\title{Create a set of scores}
\usage{
apply_scores(df, scores)
}
\arguments{
\item{df}{A data frame to score.}

\item{scores}{A data frame of score specifications. See \link{scores_init}.}
}
\value{
\code{df}, with scores added as columns.
}
\description{
Apply a table of score specification, where each specification is represented
as a row, to a data frame. The set of created scores are then added back to
the original data frame as columns.
}
\examples{
data <- tibble::tibble(
  x = 1:10
)

scores <- create_score(
  scores_init, score_type = "Linear", colname = "x", score_name = "Default",
  weight = 1, lb = 1, ub = 6, exponential = FALSE
)

apply_scores(data, scores)

}
\seealso{
\itemize{
\item \link{scores_init} for the format of the \code{scores} table.
\item \code{\link[=create_score]{create_score()}} and \code{\link[=delete_scores]{delete_scores()}} to create, edit and delete score
specifications.
\item \code{\link[=score_final]{score_final()}} to create a final score after these scores have been
created.
}
}
