% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{col_summary}
\alias{col_summary}
\title{Generate a summary of a column}
\usage{
col_summary(x)
}
\arguments{
\item{x}{The column to summarise, in vector form.}
}
\value{
A string: the summary of the column.
}
\description{
Generate a string summary of a numeric column. Shows important information
such as the mean and standard deviation.
}
\examples{
col_summary(1:100)

}
