% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_modules.R
\name{col_summary_module}
\alias{col_summary_module}
\alias{col_summary_ui}
\alias{col_summary_server}
\title{View a text summary of a column}
\usage{
col_summary_ui(id)

col_summary_server(id, column)
}
\arguments{
\item{id}{The namespace of the module.}

\item{column}{The column to summarise, in vector form.}
}
\description{
A shiny module that displays a text summary of a selected numeric column.
}
\seealso{
\code{\link[=col_summary]{col_summary()}}
}
