% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_plots.R
\name{correlation_plot}
\alias{correlation_plot}
\title{See the measure of correlation between each column in a data frame}
\usage{
correlation_plot(data, show_text = FALSE)
}
\arguments{
\item{data}{The data to plot}

\item{show_text}{Whether to show the correlation number over each panel in
the heatmap.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Create a heatmap of the correlation matrix of a data frame, created using
\code{\link[=cor]{cor()}}.
}
\examples{
correlation_plot(mtcars)
correlation_plot(mtcars, show_text = TRUE)

}
