% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_plot_modules.R
\name{correlation_plot_module}
\alias{correlation_plot_module}
\alias{correlation_plot_ui}
\alias{correlation_plot_server}
\title{Decide whether to show text in a correlation heatmap plot}
\usage{
correlation_plot_ui(id)

correlation_plot_server(id)
}
\arguments{
\item{id}{The namespace of the module.}
}
\value{
The server returns the user's input.
}
\description{
A shiny module that allows the user to decide the value of the \code{show_text}
argument to \code{\link[=correlation_plot]{correlation_plot()}}.
}
\seealso{
\code{\link[=correlation_plot]{correlation_plot()}}
}
