% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_filter_module.R
\name{create_filter_module}
\alias{create_filter_module}
\alias{create_filter_ui}
\alias{create_filter_server}
\title{Create a filter}
\usage{
create_filter_ui(id)

create_filter_server(id, data)
}
\arguments{
\item{id}{The namespace of the module.}

\item{data}{The data that is being filtered.}
}
\value{
The server alerts its parent of the value of the select input whenever the
button is pressed. Note that the logic for adding filters is not contained
within this module.
}
\description{
A shiny module containing a simple select input/action button combination.
The column to filter is selected, and the button is pressed to actually
create the filter.
}
\seealso{
\code{\link[=add_filter]{add_filter()}}
}
