% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_score.R
\name{create_score}
\alias{create_score}
\title{Create or edit a score specification}
\usage{
create_score(
  scores,
  editing = NA,
  score_type = NULL,
  colname = NULL,
  score_name = NULL,
  weight = NULL,
  lb = NULL,
  ub = NULL,
  centre = NULL,
  inverse = NULL,
  exponential = NULL,
  logarithmic = NULL,
  magnitude = NULL,
  custom_args = NULL
)
}
\arguments{
\item{scores}{A data frame of score specifications. See \link{scores_init}.}

\item{editing}{The row number of the score specification to edit. If NA (the
default), add the newly created score to the end of \code{scores}.}

\item{score_type}{The method to use when creating the score. Either "Linear",
"Peak" or "Custom coordinates".}

\item{colname}{The name of the column to score.}

\item{score_name}{The resulting name of the score. If "Default", the score
will be given a sensible and informative default name.}

\item{weight}{The weight of the score when calculating the final score. See
\code{\link[=score_final]{score_final()}}.}

\item{lb}{For a linear or peak score, the lower bound of the score. This will
produce a score of 0 in most cases.}

\item{ub}{For a linear or peak score, the upper bound of the score. This will
usually produce a score of 1 for linear scores and a score of 0 for peak
scores.}

\item{centre}{For a peak score, the centre of the score. This will produce
a score of 1 unless inverse is \code{TRUE}.}

\item{inverse}{For a peak score, whether to invert the scoring method.}

\item{exponential}{Whether to apply an exponential transformation to the
resulting score.}

\item{logarithmic}{If \code{exponential} is \code{TRUE}, whether to invert the
exponential transformation.}

\item{magnitude}{If \code{exponential} is \code{TRUE}, the numeric magnitude of the
exponential transformation.}

\item{custom_args}{For a custom score, a data frame of coordinates used to
create the score.}
}
\value{
If any required arguments are not valid, \code{scores} is returned as is.
If \code{editing} is \code{NA}, \code{scores} is returned with an ectra row.
If \code{editing} is a valid row number, \code{scores} with 1 edited row is returned.
}
\description{
Add to a table of score specifications, where each one is represented by a
row. The score specification will only be added if all arguments are valid.
Alternatively, edit an existing score specification with the \code{editing}
parameter.
}
\section{Linear scores}{
When the \code{score_name} is "Linear", a linear score is created. To make the
score you need to specify the \code{lb} and \code{ub} arguments.

If the column value is less than or equal to the \code{lb} argument, the score is
0.
If the column value is more than or equal to the \code{ub} argument, the score is
1.

Otherwise, the score is defined is the proportion of the distance of the
column value between the \code{lb} and \code{ub}.

If the \code{lb} argument is more than the \code{ub} argument, the score is inverted.
This means that the \code{lb} produces a score of 1, the \code{ub} produces a score of
0, etc.
}

\section{Peak scores}{
When the \code{score_name} is "Peak", a peak score is created. To make the score
you need to specify the \code{lb}, \code{ub}, \code{centre} and \code{inverse} arguments. The
\code{lb}, \code{ub} and \code{centre} arguments must be numeric, and the \code{centre} must be
between the \code{lb} and \code{ub}.

If the column value is less than or equal to the \code{lb} argument, the score is
0.
If the column value is equal to the \code{centre} argument, the score is 1.
If the column value is more than or equal to the \code{ub} argument, the score is
1.

If the column value is in between the \code{lb} and \code{centre} arguments, the score
is defined as the proportion of the column value along between the \code{lb} and
\code{centre}.
If the column value is in between the \code{centre} and \code{ub} arguments, the score
is defined as the proportion of the column value along between the \code{ub} and
\code{centre}.

When \code{inverse} is \code{TRUE}, the score is inverted: the lower bound and upper
bound produce a score of 1, and the centre produces a score of 0.
}

\section{Custom scores}{
When \code{score_type} is "Custom coordinates", a custom score is created. This
allows you to define a set of coordinates, where the x coordinate is a value
in the column, and the y coordinate is a score between 0 and 1. The score
will then be created by connecting the coordinates together. The coordinates
should be in the form of a data frame, with the x coordinates in the 'x'
column and the y coordinates in the 'y' column.

This can be used to create a huge variety of different scores.

For more detailed information, along with plots that demonstrate the 3
scoring methods, see the Prototype 2 vignette:
\code{vignette("prototype2", package = "financialDataAnalysis")}
}

\examples{
scores <- create_score(
  scores_init, score_type = "Linear", colname = "x", score_name = "Default",
  weight = 1, lb = 1, ub = 6, exponential = FALSE
)

# Change the weight
create_score(
  scores, editing = 1, score_type = "Linear", colname = "x", 
  score_name = "Default", weight = 2, lb = 1, ub = 6, exponential = FALSE
)

}
\seealso{
\itemize{
\item \link{scores_init} for the format of the \code{scores} table.
\item \code{\link[=delete_scores]{delete_scores()}} for the inverse operation.
\item \code{\link[=apply_scores]{apply_scores()}} for the creation of the actual scores.
}
}
