% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_plot.R
\name{custom_plot}
\alias{custom_plot}
\title{Create a customisable plot}
\usage{
custom_plot(df, plotting_method, ...)
}
\arguments{
\item{df}{The data to plot.}

\item{plotting_method}{The type of plot to create (either "line", "scatter"
or "histogram")}

\item{...}{A set of arguments mapping certain variables to aesthetics. Each
argument should be specified in the format \code{aesthetic = "column_name"}, where
\code{aesthetic} is a visual property that a variable can be mapped to (e.g. x,
colour), and \code{column_name} is the name of a column in your data.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Create a plot which is fully customisable by the user.
}
\section{Line graphs}{
Create line graphs passing in "line" to the \code{plotting_method} argument.

Line graphs accept the following aesthetics:
\itemize{
\item \code{x} - the variable on the x axis.
\item \code{y} - the variable on the y axis.
\item \code{colour} - the colour of the line.
}

\code{x} and \code{y} are required arguments, meaning that they must be supplied for a
plot to be outputted.
}

\section{Scatter graphs}{
Create scatter graphs passing in "scatter" to the \code{plotting_method} argument.

Scatter graphs accept the following aesthetics:
\itemize{
\item \code{x} - the variable on the x axis.
\item \code{y} - the variable on the y axis.
\item \code{colour} - the colour of the point.
\item \code{size} - the size of the point.
\item \code{shape} - the shape of the point.
}

\code{x} and \code{y} are required arguments, meaning that they must be supplied for a
plot to be outputted.
}

\section{Histograms}{
Create histograms passing in "histogram" to the \code{plotting_method} argument.

Histograms accept the following aesthetics:
\itemize{
\item \code{x} - the variable on the x axis.
\item \code{colour} - the colour of the bar.
\item \code{size} - the size of the bar.
}

\code{x} is a required arguments, meaning that it must be supplied for a plot to
be outputted.

The y aesthetic of a histogram is the frequency density of the x coordinate.
}

\examples{
data <- tibble::tibble(
  a = 1:10,
  b = 10:1
)

custom_plot(data, "line", x = "a", y = "b")

}
\seealso{
\code{\link[=print_plot]{print_plot()}}
}
