% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_plot_module.R
\name{custom_plot_module}
\alias{custom_plot_module}
\alias{custom_plot_ui}
\alias{custom_plot_server}
\title{Input the arguments for a custom plot}
\usage{
custom_plot_ui(id)

custom_plot_server(id, custom, data)
}
\arguments{
\item{id}{The namespace of the module.}

\item{custom}{Whether the user wants to make a custom plot.}

\item{data}{The data that is being plotted.}
}
\value{
The server returns a list of validated arguments, along with the inputted
plot type.
}
\description{
A shiny module that allows the user to input the arguments in order to create
a custom plot.
}
\seealso{
\itemize{
\item \code{\link[=custom_plot]{custom_plot()}}
\item The modules for the different custom plot types: \link{custom_plot_modules}
}
}
