% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_plot_module.R
\name{custom_plot_modules}
\alias{custom_plot_modules}
\alias{line_graph_ui}
\alias{line_graph_server}
\alias{scatter_graph_ui}
\alias{scatter_graph_server}
\alias{histogram_ui}
\alias{histogram_server}
\title{Input arguments for different custom plots}
\usage{
line_graph_ui(id)

line_graph_server(id, data)

scatter_graph_ui(id)

scatter_graph_server(id, data)

histogram_ui(id)

histogram_server(id, data)
}
\arguments{
\item{id}{The namespace of the module.}

\item{data}{The data that is being plotted}
}
\value{
Each module's server returns a list of validated arguments.
}
\description{
A set of shiny modules that allow the user to input a set of arguments
depending on the plot type.
}
\seealso{
\link{custom_plot_module}
}
