% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_score_module.R
\name{custom_row_module}
\alias{custom_row_module}
\alias{custom_row_ui}
\alias{custom_row_server}
\title{Input a set of coordinates for a custom score}
\usage{
custom_row_ui(id, n, x = 0, y = 0)

custom_row_server(id)
}
\arguments{
\item{id}{The namespace of the module.}

\item{n}{The row number - unique for every row.}

\item{x}{The initial value of the x coordinate.}

\item{y}{The initial value of the y coordinate.}
}
\value{
The server returns a \code{\link[tibble:tibble]{tibble::tibble_row()}} containing the x and y
coordinates.
}
\description{
A shiny module that contains the UI and server logic for a single row (or a
single set of coordinates) of a custom score.
}
\details{
Note that the row number (\code{n}) may not reflect the actual position of the row
within the app. As an example, consider three rows (\code{n} is 1, 2 and 3
respectively). If the second row is deleted, since the UI is not changed, the
row which will now be second will have a row number of 3.
}
\seealso{
\link{custom_score_module}
}
