% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_score_module.R
\name{custom_score_module}
\alias{custom_score_module}
\alias{custom_score_ui}
\alias{custom_score_server}
\title{Input custom coordinates to create a custom score}
\usage{
custom_score_ui(id)

custom_score_server(id, column, reset, editing_row)
}
\arguments{
\item{id}{The namespace of the module.}

\item{column}{The column currently being scored.}

\item{reset}{When this value is changed, all inputs will reset.}

\item{editing_row}{The row that the user is currently editing. When this
changes, all inputs will be updated to the state of the score that is being
edited.}
}
\value{
The server returns a validated \code{\link[tibble:tibble]{tibble::tibble()}} of coordinates.
}
\description{
A shiny module that allows the user to specify the custom coordinates for a
custom score.
}
\details{
The module's initial state uses two coordinates to create a linear score.
}
\seealso{
\link{custom_row_module}
}
