% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input_module.R
\name{data_input_module}
\alias{data_input_module}
\alias{data_input_ui}
\alias{data_input_server}
\title{Upload data}
\usage{
data_input_ui(id)

data_input_server(id)
}
\arguments{
\item{id}{The namespace of the module.}
}
\value{
The server returns the data frame that has been uploaded (or the default
stock data).
}
\description{
A shiny module that contains the Data Input page, allowing the user to
upload files. If no valid files are uploaded, the default stock data is
returned.
}
\seealso{
\code{\link[=input_data]{input_data()}}
}
