% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_module.R
\name{data_module}
\alias{data_module}
\alias{data_ui}
\alias{data_server}
\title{View a data frame.}
\usage{
data_ui(id)

data_server(id, data, filters)
}
\arguments{
\item{id}{The namespace of the module.}

\item{data}{The data to be viewed, filtered and sorted.}

\item{filters}{A table of filters to apply to the data.}
}
\description{
A shiny module that allows the user to view some data, and download it in
various file formats. The user can choose which columns are displayed,
although the number of columns are limited to 15.
}
\details{
The data is displayed using \code{\link[reactable:reactable]{reactable::reactable()}}, and the columns are
sortable and filterable.
}
