% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_stock_data}
\alias{default_stock_data}
\title{A large table of stock data}
\format{
A tibble with 497 rows and 100 variables.
}
\usage{
default_stock_data
}
\description{
A large amount of data fetched from 'iex.cloud' about almost 500 stocks.
Each row represents a stock.
}
\examples{
default_stock_data

}
\keyword{datasets}
