% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_scores.R
\name{delete_scores}
\alias{delete_scores}
\title{Remove one or more score specifications}
\usage{
delete_scores(scores, rows)
}
\arguments{
\item{scores}{A data frame of score specifications. See \link{scores_init}.}

\item{rows}{A numeric vector of rows to remove from the table.}
}
\value{
\code{scores} with some rows removed.
}
\description{
Delete one or more rows from a scores table, where each row represents a
score specification.
}
\examples{
scores <- create_score(
  scores_init, score_type = "Linear", colname = "x", score_name = "Default",
  weight = 1, lb = 1, ub = 6, exponential = FALSE
)

delete_scores(scores, 1)

}
\seealso{
\itemize{
\item \link{scores_init} for the format of the \code{scores} table.
\item \code{\link[=create_score]{create_score()}} for the inverse operation.
\item \code{\link[=apply_scores]{apply_scores()}} for the creation of the actual scores.
}
}
