% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_df.R
\name{download_df}
\alias{download_df}
\title{Write a data frame to a file}
\usage{
download_df(df, file_type, file)
}
\arguments{
\item{df}{The data frame to write.}

\item{file_type}{The type of the file to create, either "CSV" or "Excel".}

\item{file}{A string. The file path to write to.}
}
\value{
The file path that was written to.
}
\description{
Take a data frame and write it to a CSV (.csv) or Excel (.xlsx) file.
}
\examples{
file <- tempfile(fileext = ".csv")
data <- tibble::tibble(
  x = 1:10
)

download_df(data, "CSV", file)

}
