% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing_module.R
\name{editing_module}
\alias{editing_module}
\alias{editing_ui}
\alias{editing_server}
\title{Display which score the user is editing}
\usage{
editing_ui(id)

editing_server(id, editing, scores)
}
\arguments{
\item{id}{The namespace of the module.}

\item{editing}{The row number of the row currently being edited.}

\item{scores}{The table of scores.}
}
\value{
The server alerts its parent when the user indicates that they want to stop
editing.
}
\description{
A shiny module that displays the name of the score that the user is editing
}
