% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favourite_stock.R
\name{favourite_stock}
\alias{favourite_stock}
\title{Mark a stock as a 'favourite'}
\usage{
favourite_stock(df, x)
}
\arguments{
\item{df}{The data frame}

\item{x}{The row to 'favourite'}
}
\value{
A modified version of \code{df}.
}
\description{
Switch the 'favourite' column for a specified row. This feature is no longer
included in the final application.
}
\examples{
data <- tibble::tibble(
  x = 1:10,
  favourite = FALSE
)

favourite_stock(data, 5)

}
