% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{fDA_dependencies}
\alias{fDA_dependencies}
\title{Dependencies}
\usage{
fDA_dependencies()
}
\value{
A collection of HTML tags.
}
\description{
Get a list of HTML dependencies to be embedded in the app. This includes the
internal CSS and JavaScript files specific to this app, along with any other
external HTML dependencies.
}
