% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters_module.R
\name{filter_module}
\alias{filter_module}
\alias{filter_ui}
\alias{filter_server}
\title{Edit a single numeric filter}
\usage{
filter_ui(id, n, name, column)

filter_server(id)
}
\arguments{
\item{id}{The namespace of the module.}

\item{n}{The filter number - unique for each filter.}

\item{name}{The column name that is being filtered.}

\item{column}{The column that is being filtered.}
}
\value{
The server returns a \code{\link[tibble:tibble]{tibble::tibble_row()}} containing the min and max of the
filter.
}
\description{
A shiny module that allows the user to edit a single numeric filter. The
filter consists of a slider range input and two numeric inputs, where the
numeric inputs are linked to the values of the slider input.
}
\details{
The link between the numeric and slider inputs means that when the numeric
inputs are updated, the slider inputs will update, and vice versa.

Note that the filter number (\code{n}) may not reflect the actual position of the
filter within the app. As an example, consider three filters (\code{n} is 1, 2 and
3 respectively). If the second filter is deleted, since the UI is not
changed, the filter which will now be second will have a filter number of 3.
}
\seealso{
\link{filters_module}
}
