% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{filters_init}
\alias{filters_init}
\title{A table of filters}
\format{
A tibble with 0 rows and 5 variables
\describe{
\item{type}{The filter type, either "character" or "numeric"}
\item{colname}{The name of the column to filter}
\item{pattern}{For character filters, the pattern to match to the column
value}
\item{min}{For numeric filters, the minimum value of the column}
\item{max}{For numeric filters, the maximum value of the column}
}
}
\usage{
filters_init
}
\description{
The initial state of a data frame, where each row represents a filter.
}
\details{
Character filters work on columns that contain strings (sequences of
letters). For each row, if the value of the column contains the \code{pattern}
of the filter, it is kept in the filtered data frame. Otherwise, the row is
filtered out.

For example, if the pattern was "at", and the column values were "cat",
"match" and "bit", then the first two values would be kept in the filtered
data frame since they contain the pattern.

Numeric filters work on columns containing numbers. For each row, if the
column value is between the \code{min} value and the \code{max} value, it is kept in
the filtered data frame. Otherwise, it is removed.
}
\examples{
filters_init

}
\seealso{
\itemize{
\item \code{\link[=add_filter]{add_filter()}}, \code{\link[=edit_filter]{edit_filter()}} and \code{\link[=remove_filter]{remove_filter()}} to manipulate the
filters table.
\item \code{\link[=apply_filters]{apply_filters()}} to apply the filters table you have created to some data.
}
}
\keyword{datasets}
