% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters_module.R
\name{filters_module}
\alias{filters_module}
\alias{filters_ui}
\alias{filters_server}
\title{Edit a set of numeric filters}
\usage{
filters_ui(id)

filters_server(id, data, add)
}
\arguments{
\item{id}{The namespace of the module.}

\item{data}{The data to filter.}

\item{add}{The input (from another module) that signifies the creation of a
new filter.}
}
\value{
The server returns a \code{\link[tibble:tibble]{tibble::tibble()}} of filters.
}
\description{
A shiny module that allows the user to edit a set of numeric filters for a
data frame.
}
\seealso{
\code{\link[=apply_filters]{apply_filters()}} \link{filter_module}
}
