% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{financialDataAnalysis}
\alias{financialDataAnalysis}
\title{Run the 'financialDataAnalysis' app}
\usage{
financialDataAnalysis(...)
}
\arguments{
\item{...}{Passed into \code{\link[shiny:shinyApp]{shiny::shinyApp()}}}
}
\value{
An object that represents the app.
}
\description{
Run the project as a Shiny application. The function calls
\code{\link[shiny:shinyApp]{shiny::shinyApp()}} on the UI and server functions.
}
\examples{
\dontrun{
  financialDataAnalysis()
}

}
\seealso{
\code{\link[=ui]{ui()}} \code{\link[=server]{server()}}
}
