% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_price_module.R
\name{forecast_price_module}
\alias{forecast_price_module}
\alias{forecast_price_ui}
\alias{forecast_price_server}
\title{Select and forecast stocks}
\usage{
forecast_price_ui(id)

forecast_price_server(id)
}
\arguments{
\item{id}{The namespace of the module.}
}
\description{
A shiny module which contains the Forecast Price page. Allows the user to
search for stocks, see details about selected stocks and forecast the price
of a stock over time.
}
