% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lightgbm_model.R
\name{get_lightgbm_model}
\alias{get_lightgbm_model}
\title{Model the residuals of stock price predictions.}
\usage{
get_lightgbm_model(x)
}
\arguments{
\item{x}{Which model to get: either "daily" or "monthly".}
}
\value{
A fitted \code{\link[workflows:workflow]{workflows::workflow()}} object.
}
\description{
Get the daily or monthly LightGBM models, which are trained to predict the
residuals of the Prophet model.
}
\details{
This model is designed to perform iterative forecasting. This means that it
can predict residuals for a certain horizon, then assumes these predictions
as truth and uses lagged values as predictors, and makes new predictions,
and so on.

The process to get the models is complex because \code{\link[=readRDS]{readRDS()}} does not work
for LightGBM models.
}
\examples{
get_lightgbm_model("daily")

}
\seealso{
\link{prophet_models} \code{\link[=predict_price]{predict_price()}}
}
