% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scores.R
\name{get_scores}
\alias{get_scores}
\title{Get the created scores from a scored data frame.}
\usage{
get_scores(df, scores, final_score = TRUE)
}
\arguments{
\item{df}{The scored data frame.}

\item{scores}{A data frame of score specifications. See \link{scores_init}.}

\item{final_score}{Whether the data frame has been given a final score by
\code{\link[=score_final]{score_final()}}.}
}
\value{
A data frame of scores.
}
\description{
Extract the score columns from a data frame scored using \code{\link[=apply_scores]{apply_scores()}}.
}
\examples{
data <- tibble::tibble(
  x = 1:10
)

scores <- create_score(
  scores_init,
  score_type = "Linear", colname = "x", score_name = "Default",
  weight = 1, lb = 1, ub = 6, exponential = FALSE
)
scores <- create_score(
  scores,
  score_type = "Peak", colname = "x", score_name = "Peak score",
  weight = 2, lb = 2, ub = 8, centre = 5, inverse = FALSE,
  exponential = FALSE
)

scored <- apply_scores(data, scores)
scored <- score_final(scored, scores)

get_scores(scored, scores)

}
