% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{input_data}
\alias{input_data}
\title{Convert a set of files into a data frame}
\usage{
input_data(files, default_data = default_stock_data, combine = FALSE)
}
\arguments{
\item{files}{A vector of file paths. The function only accepts CSV and Excel
files.}

\item{default_data}{The default data, included for testing purposes. It is
not recommended to change this from its default.}

\item{combine}{If \code{TRUE}, the resulting data frame will be combined with the
default data frame.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Read one or more files and combine them into a single data frame. If the
result is invalid, the default stock data is returned
}
\examples{
file <- tempfile(fileext = "csv")
data <- tibble::tibble(
  x = 1:100
)
readr::write_csv(data, file)

input_data(file)

}
\seealso{
\link{default_stock_data}
}
