% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_input_modules.R
\name{linear_score_module}
\alias{linear_score_module}
\alias{linear_score_ui}
\alias{linear_score_server}
\title{Input the arguments specific to a linear score.}
\usage{
linear_score_ui(id)

linear_score_server(id, column, reset, editing_row)
}
\arguments{
\item{id}{The namespace of the module.}

\item{column}{The column currently being scored.}

\item{reset}{When this value is changed, all inputs will reset.}

\item{editing_row}{The row that the user is currently editing. When this
changes, all inputs will be updated to the state of the score that is being
edited.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble_row()}} containing the two inputs, or NULL if any are
invalid.
}
\description{
A shiny module that allows the user to input the lower bound and upper bound
of a score.
}
