% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{method_fun}
\alias{method_fun}
\title{\if{html}{\out{<div class="sourceCode">}}\preformatted{method_fun <- 
  list(
    line = "geom_line",
    scatter = "geom_point",
    histogram = "geom_histogram"
  )
}\if{html}{\out{</div>}}}
\format{
An object of class \code{list} of length 3.
}
\usage{
method_fun
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{method_fun <- 
  list(
    line = "geom_line",
    scatter = "geom_point",
    histogram = "geom_histogram"
  )
}\if{html}{\out{</div>}}
}
\keyword{datasets}
