% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_input_modules.R
\name{peak_score_module}
\alias{peak_score_module}
\alias{peak_score_ui}
\alias{peak_score_server}
\title{Input the arguments specific to a peak score}
\usage{
peak_score_ui(id)

peak_score_server(id, column, reset, editing_row)
}
\arguments{
\item{id}{The namespace of the module.}

\item{column}{The column currently being scored.}

\item{reset}{When this value is changed, all inputs will reset.}

\item{editing_row}{The row that the user is currently editing. When this
changes, all inputs will be updated to the state of the score that is being
edited.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble_row()}} containing the four inputs, or NULL if any are
invalid.
}
\description{
A shiny module that allows the user to the input the lower bound, upper
bound, centre and inverse arguments to a peak score.
}
