% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_module.R
\name{plot_data_module}
\alias{plot_data_module}
\alias{plot_data_ui}
\alias{plot_data_server}
\title{Create a variety of plots}
\usage{
plot_data_ui(id)

plot_data_server(id, data, scores, interactive)
}
\arguments{
\item{id}{The namespace of the module.}

\item{data}{The data to plot.}

\item{scores}{The data frame of score specifications created by the user. See
\link{scores_init}.}

\item{interactive}{Whether any plots created should be interactive.}
}
\description{
A shiny module that contains the Plot Data page, which allows the user to
create a variety of default or custom plots.
}
