% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_module.R
\name{plot_module}
\alias{plot_module}
\alias{plot_ui}
\alias{plot_server}
\title{Create and display a plot}
\usage{
plot_ui(id)

plot_server(
  id,
  type,
  data,
  scores,
  performance_col,
  custom_type,
  custom_args,
  create
)
}
\arguments{
\item{id}{The namespace of the module.}

\item{type}{The plot type.}

\item{data}{The data to plot.}

\item{scores}{The data frame of score specifications created by the user. See
\link{scores_init}.}

\item{performance_col}{If \code{type} is "Score performance", the column to plot
against the scores.}

\item{custom_type}{If \code{type} is "Custom", the plotting method of the custom
plot.}

\item{custom_args}{If \code{type} is "Custom", the arguments to create the custom
plot.}

\item{create}{The button input that specifies that a plot should be
generated.}
}
\description{
A shiny module which creates and shows the plot that the user has specified.
}
