% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{plot_predictions}
\alias{plot_predictions}
\title{Plot price predictions}
\usage{
plot_predictions(predicted)
}
\arguments{
\item{predicted}{A data frame of predictions, likely generated by the
\code{\link[=predict_price]{predict_price()}} function.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Take a set of predictions and generate a line plot of them.
}
\seealso{
\code{\link[=predict_price]{predict_price()}}
}
