% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_price}
\alias{predict_price}
\title{Predict the price of a stock over time}
\usage{
predict_price(
  stock,
  start_date = lubridate::today(),
  end_date = lubridate::today() + lubridate::days(60),
  freq = c("daily", "monthly"),
  hostess = dummy_hostess()
)
}
\arguments{
\item{stock}{The stock to generate predictions for (e.g. "GOOGL").}

\item{start_date}{The first date of the predictions.}

\item{end_date}{The final date of the predictions.}

\item{freq}{the frequency of the predictions.}

\item{hostess}{A loader to show the progress of the predictions. A
\link[waiter:hostess]{waiter::Hostess} object.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} of stock predictions.
}
\description{
Generate predictions for the price of a stock over a specified time period.
Choose between daily or monthly predictions.
}
\details{
The prediction process involves using a Prophet model to generate
predictions, then using a lightgbm model to predict the residuals and adding
the two together.
}
\seealso{
\itemize{
\item \code{\link[=plot_predictions]{plot_predictions()}} to plot your generated predictions.
\item \link{prophet_models} and \code{\link[=get_lightgbm_model]{get_lightgbm_model()}} for the underlying models.
}
}
