% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_price_module.R
\name{predict_price_module}
\alias{predict_price_module}
\alias{predict_price_ui}
\alias{predict_price_server}
\title{Generate stock price predictions}
\usage{
predict_price_ui(id)

predict_price_server(id, stock, interactive)
}
\arguments{
\item{id}{The namespace of the module.}

\item{stock}{The selected stock to generate predictions for.}

\item{interactive}{Whether the created plot should be interactive.}
}
\description{
Allows the user to specify a start date and an end date, then predicts the
price of a stock over the specified period, and plots the results.
}
