% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_plot.R
\name{print_plot}
\alias{print_plot}
\title{Safely print a plot}
\usage{
print_plot(plot)
}
\arguments{
\item{plot}{The plot to print.}
}
\value{
\code{plot}
}
\description{
Suppress all errors and warnings while printing a plot.
}
\seealso{
\code{\link[=custom_plot]{custom_plot()}}
}
