% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daily_prophet_model}
\alias{daily_prophet_model}
\title{Predict the price of a stock over time}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 501 rows and 3 columns.
}
\usage{
daily_prophet_model
}
\description{
Use a Prophet model to predict the price of a stock over time.
}
\keyword{datasets}
