% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prophet_models}
\alias{prophet_models}
\alias{daily_prophet_model}
\alias{monthly_prophet_model}
\title{Predict the price of a stock over time}
\format{
A tibble with 491 / 470 rows and 3 variables.
\describe{
\item{ticker}{The ticker that the model was fitted on}
\item{data}{The training data that the model was fitted on}
\item{fit}{The fitted Prophet model object}
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 482 rows and 3 columns.
}
\usage{
daily_prophet_model

monthly_prophet_model
}
\description{
Use a Prophet model to predict the price of a stock over time. Use
\code{daily_prophet_model} for daily predictions, and \code{monthly_prophet_model} for
monthly predictions.
}
\details{
Each model was fitted on the time series data for an individual stock.
}
\examples{
daily_prophet_model
monthly_prophet_model

}
\seealso{
\code{\link[=predict_price]{predict_price()}} \code{\link[=get_lightgbm_model]{get_lightgbm_model()}}
}
\keyword{datasets}
