% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_filter.R
\name{remove_filter}
\alias{remove_filter}
\title{Remove a filter from a table of filters}
\usage{
remove_filter(filters, x)
}
\arguments{
\item{filters}{A data frame of filters. See \link{filters_init}.}

\item{x}{A number specifying the filter number to remove (indexed at 1).}
}
\value{
\code{filters} with one less row.
}
\description{
Delete a filter for a data frame, by removing a row from a data frame, where
each row represents a filter.
}
\examples{
data <- tibble::tibble(
  x = 1:10,
  y = "a"
)

filters <- add_filter(filters_init, "x", data)
remove_filter(filters, 1)

}
\seealso{
\itemize{
\item \link{filters_init} for the filters table.
\item \code{\link[=add_filter]{add_filter()}} and \code{\link[=edit_filter]{edit_filter()}} to further edit the filters table.
\item \code{\link[=apply_filters]{apply_filters()}} to apply the filters you have created to your data.
}
}
