% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_plots.R
\name{score_distributions}
\alias{score_distributions}
\title{Plot the spread of a set of scores}
\usage{
score_distributions(scores)
}
\arguments{
\item{scores}{A data frame containing a set of numerical scores. If your data
frame has been scored using \code{\link[=apply_scores]{apply_scores()}}, use \code{\link[=get_scores]{get_scores()}} to extract
the actual scores.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Use a boxplot and jitter plot to display the spread and arrangement of a set
of scores. Can often be used to view non-useful scores, where the majority
of the scores are the same.
}
\details{
Scores must be between 0 and 1.
}
\examples{
data <- tibble::tibble(
  x = 1:10
)

scores <- create_score(
  scores_init,
  score_type = "Linear", colname = "x", score_name = "Default",
  weight = 1, lb = 1, ub = 6, exponential = FALSE
)
scores <- create_score(
  scores,
  score_type = "Peak", colname = "x", score_name = "Peak score",
  weight = 2, lb = 2, ub = 8, centre = 5, inverse = FALSE,
  exponential = FALSE
)

scored <- apply_scores(data, scores)
score_distributions(get_scores(scored, scores, final_score = FALSE))

}
\seealso{
\code{\link[=score_performance]{score_performance()}}
}
