% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_final.R
\name{score_final}
\alias{score_final}
\title{Calculate a final score}
\usage{
score_final(df, scores)
}
\arguments{
\item{df}{A data frame. The result of \code{\link[=apply_scores]{apply_scores()}}.}

\item{scores}{scores A data frame of score specifications. See \link{scores_init}.
This must be the same as the scores passed into \code{\link[=apply_scores]{apply_scores()}}.}
}
\description{
Add a 'final_score' column to a data frame that has been scored using
\code{\link[=apply_scores]{apply_scores()}} by calculating a weighted mean of all other scores.
}
\details{
If no scores were created, the final score will not be added.]
}
\examples{
data <- tibble::tibble(
  x = 1:10
)

scores <- create_score(
  scores_init,
  score_type = "Linear", colname = "x", score_name = "Default",
  weight = 1, lb = 1, ub = 6, exponential = FALSE
)
scores <- create_score(
  scores,
  score_type = "Peak", colname = "x", score_name = "Peak score",
  weight = 2, lb = 2, ub = 8, centre = 5, inverse = FALSE,
  exponential = FALSE
)

scored <- apply_scores(data, scores)
score_final(scored, scores)

}
\seealso{
\itemize{
\item \link{scores_init} for the format of the \code{scores} table.
\item \code{\link[=apply_scores]{apply_scores()}} for the prior step.
}
}
