% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_plots.R
\name{score_performance}
\alias{score_performance}
\title{Compare a set of scores to a column}
\usage{
score_performance(df, colname, scores)
}
\arguments{
\item{df}{A data frame to be plotted.}

\item{colname}{The name of the column to plot.}

\item{scores}{A data frame containing a set of numerical scores. If your data
frame has been scored using \code{\link[=apply_scores]{apply_scores()}}, use \code{\link[=get_scores]{get_scores()}} to extract
the actual scores.}
}
\description{
Plot a line graph of a set of scores against a single column in a data frame.
Can be used to evaluate whether a score is useful in determining some other
factor.
}
\examples{
data <- tibble::tibble(
  x = 1:10
)

scores <- create_score(
  scores_init,
  score_type = "Linear", colname = "x", score_name = "Default",
  weight = 1, lb = 1, ub = 6, exponential = FALSE
)
scores <- create_score(
  scores,
  score_type = "Peak", colname = "x", score_name = "Peak score",
  weight = 2, lb = 2, ub = 8, centre = 5, inverse = FALSE,
  exponential = FALSE
)

scored <- apply_scores(data, scores)
score_performance(
  scored,
  colname = "x",
  scores = get_scores(scored, scores, final_score = FALSE)
)

}
\seealso{
\code{\link[=score_distributions]{score_distributions()}}
}
