% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_modules.R
\name{score_summary_module}
\alias{score_summary_module}
\alias{score_summary_ui}
\alias{score_summary_server}
\title{View a graphical summary of a score}
\usage{
score_summary_ui(id)

score_summary_server(id, column, score_spec, interactive)
}
\arguments{
\item{id}{The namespace of the module.}

\item{column}{The column being scored.}

\item{score_spec}{The score currently being created.}

\item{interactive}{Whether the created plot should be interactive.}
}
\description{
A shiny module that allows the user to view a plot of how the score they are
currently creating will look.
}
\seealso{
\code{\link[=score_summary]{score_summary()}}
}
