% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_input_modules.R
\name{score_type_module}
\alias{score_type_module}
\alias{score_type_ui}
\alias{score_type_server}
\title{Input the type of a score.}
\usage{
score_type_ui(id)

score_type_server(id, reset, editing_row)
}
\arguments{
\item{id}{The namespace of the module.}

\item{reset}{When this value is changed, all inputs will reset.}

\item{editing_row}{The row that the user is currently editing. When this
changes, all inputs will be updated to the state of the score that is being
edited.}
}
\value{
The server returns the score type, or NULL if it is invalid.
}
\description{
A shiny module that allows the user to input the score type of a score via a
select input.
}
