% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scores_init}
\alias{scores_init}
\title{\if{html}{\out{<div class="sourceCode">}}\preformatted{scores_init <-
  tibble::tibble(
    score_type = character(),
    colname = character(),
    score_name = character(),
    weight = numeric(),
    lb = numeric(),
    ub = numeric(),
    centre = numeric(),
    inverse = logical(),
    exponential = logical(),
    logarithmic = logical(),
    magnitude = numeric(),
    custom_args = list(),
    .rows = 0
  )
}\if{html}{\out{</div>}}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 0 rows and 12 columns.
}
\usage{
scores_init
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{scores_init <-
  tibble::tibble(
    score_type = character(),
    colname = character(),
    score_name = character(),
    weight = numeric(),
    lb = numeric(),
    ub = numeric(),
    centre = numeric(),
    inverse = logical(),
    exponential = logical(),
    logarithmic = logical(),
    magnitude = numeric(),
    custom_args = list(),
    .rows = 0
  )
}\if{html}{\out{</div>}}
}
\keyword{datasets}
