% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_table_module.R
\name{scores_table_module}
\alias{scores_table_module}
\alias{scores_table_ui}
\alias{scores_table_server}
\title{View the table of scores}
\usage{
scores_table_ui(id)

scores_table_server(id, scores)
}
\arguments{
\item{id}{The namespace of the module.}

\item{scores}{The data frame of score specifications created by the user. See
\link{scores_init}.}
}
\value{
The server returns the edited and deleted scores.
}
\description{
A shiny module that allows the user to view, edit and delete created score
specifications.
}
