% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_bar_module.R
\name{search_bar_module}
\alias{search_bar_module}
\alias{search_bar_ui}
\alias{search_bar_server}
\title{Search and select a stock.}
\usage{
search_bar_ui(id)

search_bar_server(id)
}
\arguments{
\item{id}{The namespace of the module}
}
\value{
The server returns the selected stock.
}
\description{
A shiny module that allows the user to search through stocks using the ticker
and name of the stock.
}
