% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_stocks.R
\name{search_stocks}
\alias{search_stocks}
\title{Search through stocks}
\usage{
search_stocks(x, pattern)
}
\arguments{
\item{x}{The data frame of stocks (e.g. the \link{default_stock_data}).}

\item{pattern}{The pattern to use to filter \code{x}.}
}
\value{
A filtered version of \code{x}
}
\description{
Filter through a data frame of stocks using a search pattern. The pattern is
split up by spaces, and each resulting pattern is matched to the values of
the stock name and ticker.
}
\examples{
search_stocks(default_stock_data, "Al p")

}
