% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Main server}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{A \code{\link[shiny:reactiveValues]{shiny::reactiveValues()}} object containing all inputs sent
from the client/UI to the server.}

\item{output}{A "shinyoutput" object containing all values sent from the
server to the client/UI.}

\item{session}{A "ShinySession" object. This is an environment that contains
information and functionality relating to the current session (see
\link[shiny:session]{shiny::session}).}
}
\description{
The main server logic for the application.
}
