% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_module.R
\name{settings_module}
\alias{settings_module}
\alias{settings_ui}
\alias{settings_server}
\title{Specify various site settings}
\usage{
settings_ui(id)

settings_server(id)
}
\arguments{
\item{id}{The namespace of the module.}
}
\value{
The user's inputted settings
}
\description{
A shiny module that contains the Settings page. Allows the user to input
various settings. At the moment, the only setting is one which allows them
to disable interactive plots.
}
