% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_df.R
\name{sort_df}
\alias{sort_df}
\title{Sort a data frame by one of its columns}
\usage{
sort_df(df, colname, desc = FALSE)
}
\arguments{
\item{df}{The data frame to sort.}

\item{colname}{The column to sort with.}

\item{desc}{Whether to sort in descending order.}
}
\value{
\code{df}, sorted.
}
\description{
Arrange a daat frame by one of its columns, either in ascending or descending
order.
}
\examples{
data <- tibble::tibble(
  x = 10:1,
  y = c(letters[6:10], letters[5:1])
)

sort_df(data, "x")
sort_df(data, "y", desc = TRUE)

}
