% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{stock_summary}
\alias{stock_summary}
\title{Create a summary of a stock}
\usage{
stock_summary(df, x)
}
\arguments{
\item{df}{The stock data. See \link{default_stock_data}.}

\item{x}{The row number of the stock.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble_row()}}.
}
\description{
Summarise a stock by showing the data corresponding to the stock.
}
\examples{
stock_summary(default_stock_data, 1)

}
