% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_modules.R
\name{stock_summary_module}
\alias{stock_summary_module}
\alias{stock_summary_ui}
\alias{stock_summary_server}
\title{View a summary of a selected stock}
\usage{
stock_summary_ui(id)

stock_summary_server(id, stock)
}
\arguments{
\item{id}{The namespace of the module.}

\item{stock}{The stock that has been selected.}
}
\description{
A shiny module that allows the user to see a tabular summary of a stock they
have selected.
}
\seealso{
\code{\link[=stock_summary]{stock_summary()}}
}
