% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{training_data}
\alias{training_data}
\alias{daily_training_data}
\alias{monthly_training_data}
\title{Stock price training data}
\format{
A tibble with 63,617 / 57,793 rows and 4 variables.
\describe{
\item{ticker}{The ticker/symbol of the stock}
\item{ref_date}{The date on which the price was recorded}
\item{price}{The price of the stock (USD)}
\item{residuals}{The residuals (difference between predictions and actual
price) of the Prophet model}
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 58056 rows and 4 columns.
}
\usage{
daily_training_data

monthly_training_data
}
\description{
A set of tables containing the stock price and the residuals of the Prophet
model over a daily or monthly period. This data was used to train the
LightGBM model.
}
\examples{
daily_training_data
monthly_training_data

}
\seealso{
\link{prophet_models} \code{\link[=get_lightgbm_model]{get_lightgbm_model()}}
}
\keyword{datasets}
