% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_input_modules.R
\name{universal_score_module}
\alias{universal_score_module}
\alias{universal_score_ui}
\alias{universal_score_server}
\title{Input the score arguments that are universal.}
\usage{
universal_score_ui(id)

universal_score_server(id, data, reset, editing_row)
}
\arguments{
\item{id}{The namespace of the module.}

\item{data}{The data to be scored.}

\item{reset}{When this value is changed, all inputs will reset.}

\item{editing_row}{The row that the user is currently editing. When this
changes, all inputs will be updated to the state of the score that is being
edited.}
}
\value{
The server returns the three inputs.
}
\description{
A shiny module that allows the user to input the column name, score name and
weight of a score specification.
}
