% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_data_module.R
\name{view_data_module}
\alias{view_data_module}
\alias{view_data_ui}
\alias{view_data_server}
\title{Filter, sort and view scored data}
\usage{
view_data_ui(id)

view_data_server(id, data, scores)
}
\arguments{
\item{id}{The namespace of the module.}

\item{data}{The data frame, after the scoring process.}

\item{scores}{The data frame of score specifications created by the user. See
\link{scores_init}.}
}
\description{
A shiny module that contains the View Data page. Allows the user to create
filters (which are then applied to the data), sort their data and download
it in a variety of formats.
}
