% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financialDataAnalysis-package.R
\docType{package}
\name{financialDataAnalysis-package}
\alias{financialDataAnalysis-package}
\alias{_PACKAGE}
\title{financialDataAnalysis: A shiny app for scoring, viewing and plotting financial data}
\description{
Score, view and plot data. The package includes a rich default dataset of stock data, but allows the user to upload their own data. A predictive model is included to forecast the price of a stock over time.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ashbythorpe/financialDataAnalysis}
  \item \url{https://ashbythorpe.github.io/financialDataAnalysis/}
  \item Report bugs at \url{https://github.com/ashbythorpe/financialDataAnalysis/issues}
}

}
\author{
\strong{Maintainer}: Ashby Thorpe \email{ashbythorpe@gmail.com} (\href{https://orcid.org/0000-0003-3106-099X}{ORCID})

}
\keyword{internal}
