% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scores_init}
\alias{scores_init}
\title{A table of score specifications}
\format{
A tibble with 0 rows and 12 variables.
\describe{
\item{score_type}{The method to use when creating the score: either
"Linear", "Peak" or "Custom coordinates"}
\item{colname}{The name of the column to score}
\item{score_name}{The resulting name of the score}
\item{weight}{The weight of the score when calculating the final score}
\item{lb}{The lower bound of the score}
\item{ub}{The upper bound of the score}
\item{centre}{The centre of the score}
\item{inverse}{Whether to invert a peak score}
\item{exponential}{Whether to apply an exponential transformation to the
score}
\item{logarithmic}{Whether to invert the exponential transformation}
\item{magnitude}{The magnitude of the exponential transformation}
\item{custom_args}{A data frame of coordinates used to create a custom
score}
}
}
\usage{
scores_init
}
\description{
The initial state of a table where each row represents a specification of a
score.
}
\examples{
scores_init

}
\seealso{
\itemize{
\item \code{\link[=create_score]{create_score()}} and \code{\link[=delete_scores]{delete_scores()}} to manipulate the table.
\item \code{\link[=apply_scores]{apply_scores()}} for the creation of the actual scores.
}
}
\keyword{datasets}
